+++ packages/iov-bns/src/bnsconnection.ts
@@ -62,6 +62,7 @@ import {
   decodeElectorate,
   decodeProposal,
   decodeToken,
+  decodeTxFeeConfiguration,
   decodeUsernameNft,
   decodeVote,
 } from "./decodeobjects";
@@ -80,6 +81,7 @@ import {
   Keyed,
   Proposal,
   Result,
+  TxFeeConfiguration,
   Validator,
   Vote,
 } from "./types";
@@ -779,6 +781,21 @@ export class BnsConnection implements AtomicSwapConnection {
     };
     const bytesToPost = bnsCodec.bytesToPost(withDummySignatures);
     return bytesToPost.length;
+  }
+
+  public async getTxFeeConfiguration(): Promise<TxFeeConfiguration | undefined> {
+    const { results } = await this.query("/", Encoding.toAscii("_c:txfee"));
+    if (results.length > 1) {
+      throw new Error(
+        `Unexpected number of results for tx fee configuration. Expected: 0/1 Got: ${results.length}`,
+      );
+    }
+    if (results.length === 0) {
+      return undefined;
+    }
+    const config = decodeTxFeeConfiguration(codecImpl.txfee.Configuration.decode(results[0].value));
+
+    return config;
   }
 
   public async getFeeQuote(transaction: UnsignedTransaction): Promise<Fee> {