+++ packages/iov-bns/types/types.d.ts
@@ -16,6 +16,10 @@ import * as codecImpl from "./generated/codecimpl";
 export interface CashConfiguration {
   readonly minimalFee: Amount | null;
 }
+export interface TxFeeConfiguration {
+  readonly freeBytes: number | null;
+  readonly baseFee: Amount | null;
+}
 /**
  * The message part of a bnsd.Tx
  *